forge_lint

Module linter

Modules

early 🔒

late 🔒

Structs

EarlyLintVisitor

Visitor struct for EarlyLintPasses

LateLintVisitor

Visitor struct for LateLintPasses

LintContext

LinterConfig

Enums

Snippet

Traits

EarlyLintPass

Trait for lints that operate directly on the AST. Its methods mirror ast::visit::Visit, with the addition of LintCotext.

LateLintPass

Trait for lints that operate on the HIR (High-level Intermediate Representation). Its methods mirror hir::visit::Visit, with the addition of LintCotext.

Lint

Linter

Trait representing a generic linter for analyzing and reporting issues in smart contract source code files. A linter can be implemented for any smart contract language supported by Foundry.