forge_lint::linter

Trait Linter

pub trait Linter: Send + Sync {
    type Language: Language;
    type Lint: Lint;

    // Required methods
    fn configure(&self, compiler: &mut Compiler);
    fn lint(&self, input: &[PathBuf], compiler: &mut Compiler);

    // Provided method
    fn init(&self) -> Compiler { ... }
}

Trait representing a generic linter for analyzing and reporting issues in smart contract source code files.

A linter can be implemented for any smart contract language supported by Foundry.

Required Associated Types

type Language: Language

The target [Language].

type Lint: Lint

The Lint type.

Required Methods

fn configure(&self, compiler: &mut Compiler)

Configure a solar [Compiler] from the given linter config.

fn lint(&self, input: &[PathBuf], compiler: &mut Compiler)

Run all lints.

The compiler should have already been configured with all the sources necessary, as well as having performed parsing and lowering.

Provided Methods

fn init(&self) -> Compiler

Build a solar [Compiler] from the given linter config.