forge_lint

Module sol

Modules

gas

high

info

med

Macros

declare_sol_lints 🔒

Macro for defining lints and relevant metadata for the Solidity linter.

Structs

AsmKeccak256

DivideBeforeMultiply

IncorrectShift

ScreamingSnakeCase

SolidityLinter

Linter implementation to analyze Solidity source code responsible for identifying vulnerabilities gas optimizations, and best practices.

StructPascalCase

VariableMixedCase

Enums

SolLint

SolLintError